ModuleRequested event support
Jint.CommonJS now supports a ModuleRequested event which is raised every
time the module loading engine is asked for a module that it can't find
in the cache.  Return a non-null JsValue from this event to support
internal modules which are dynamically fetched from instances that are
outside of Jint.CommonJS's control.